rename main.js to index.js
