--- backend/routes/middleware/mail/package.json
@@ -1,3 +1,0 @@
-{
-    "main":"./main.js"
-}
