--- backend/routes/middleware/validation/package.json
@@ -1,3 +1,0 @@
-{
-    "main":"./main.js"
-}
