+++ backend/routes/srcs/in/announcement/getAnnouncement.js
@@ -7,9 +7,9 @@ const Announcement = require('../../../Schemas/announcement')
  * @apiName GetAnnouncement
  * @apiGroup In/announcement
  * @apiDescription 拿一筆公告
- * 
+ *
  * @apiparam {String} _id (required)
- * 
+ *
  * @apiError (404) {String} description id is required/資料不存在
  * @apiError (500) {String} description 資料庫錯誤
  */
@@ -20,6 +20,6 @@ const getAnnouncement = async (req, res, next) => {
   return res.status(201).send(announcement)
 }
 
-const valid = require('../../../middleware/validation/main')
+const valid = require('../../../middleware/validation')
 const rules = [{ filename: 'required', field: '_id', method: 'get' }]
 module.exports = [valid(rules), asyncHandler(getAnnouncement)]
