+++ backend/routes/srcs/in/study/storeMatch/index.js
@@ -2,7 +2,6 @@ const asyncHandle = require('express-async-handler')
 const parseExcel = require('./parseExcel')
 const post = require('./mail')
 
-
 /**
  * @api {post} /study/sendmail 寄配對通知
  * @apiName Study_matching
@@ -15,9 +14,10 @@ const post = require('./mail')
  * 
  * @apiSuccess (203) {String[]} errors 發生錯誤的寄件者姓名 
  */
-module.exports = [parseExcel('result'),
-    asyncHandle(async (req,res,next)=>{
-        const errors = await post(__dirname+'/uploads/output.xlsx')
-        res.send({errors})
-    }
-)]
+module.exports = [
+  parseExcel('result'),
+  asyncHandle(async (req, res, next) => {
+    const errors = await post(__dirname + '/uploads/output.xlsx')
+    res.send({ errors })
+  }),
+]
